fixes DataList configuration
