+++ src/main/app/Resources/modules/content/list/components/display.jsx
@@ -1,18 +1,19 @@
 import React from 'react'
-import {PropTypes as T} from 'prop-types'
 
+import {PropTypes as T, implementPropTypes} from '#/main/app/prop-types'
 import {trans} from '#/main/app/intl/translation'
 import {Button} from '#/main/app/action/components/button'
 import {CALLBACK_BUTTON, MENU_BUTTON} from '#/main/app/buttons'
 
-import {constants} from '#/main/app/content/list/constants'
+import {DataListDisplay} from '#/main/app/content/list/prop-types'
+import DISPLAY_MODES from '#/main/app/content/list/modes'
 
 const ListDisplay = props =>
   <Button
     id="list-format-toggle"
     className="list-header-btn btn btn-text-secondary"
     type={MENU_BUTTON}
-    icon={constants.DISPLAY_MODES[props.current].icon}
+    icon={DISPLAY_MODES[props.current].icon}
     label={trans('list_display_modes_title')}
     tooltip="bottom"
     disabled={props.disabled}
@@ -21,22 +22,17 @@ const ListDisplay = props =>
       label: trans('list_display_modes'),
       items: props.available.map(display => ({
         type: CALLBACK_BUTTON,
-        icon: constants.DISPLAY_MODES[display].icon,
-        label: constants.DISPLAY_MODES[display].label,
+        icon: DISPLAY_MODES[display].icon,
+        label: DISPLAY_MODES[display].label,
         active: display === props.current,
-        callback: () => props.change(display)
+        callback: () => props.changeDisplay(display)
       }))
     }}
   />
 
-ListDisplay.propTypes = {
-  disabled: T.bool,
-  current: T.oneOf(Object.keys(constants.DISPLAY_MODES)).isRequired,
-  available: T.arrayOf(
-    T.oneOf(Object.keys(constants.DISPLAY_MODES))
-  ).isRequired,
-  change: T.func.isRequired
-}
+implementPropTypes(ListDisplay, DataListDisplay, {
+  disabled: T.bool
+})
 
 export {
   ListDisplay
