+++ src/main/app/Resources/modules/content/list/grid/constants.js
@@ -1,0 +1,12 @@
+
+const DISPLAY_TILES    = 'tiles'
+const DISPLAY_TILES_SM = 'tiles-sm'
+const DISPLAY_LIST     = 'list'
+const DISPLAY_LIST_SM  = 'list-sm'
+
+export const constants = {
+  DISPLAY_LIST_SM,
+  DISPLAY_LIST,
+  DISPLAY_TILES_SM,
+  DISPLAY_TILES
+}
