+++ src/main/app/Resources/modules/content/list/table/constants.js
@@ -1,0 +1,8 @@
+
+const DISPLAY_TABLE    = 'table'
+const DISPLAY_TABLE_SM = 'table-sm'
+
+export const constants = {
+  DISPLAY_TABLE_SM,
+  DISPLAY_TABLE
+}
