+++ src/main/app/Resources/modules/content/list/tree/constants.js
@@ -1,0 +1,6 @@
+
+const DISPLAY_TREE    = 'tree'
+
+export const constants = {
+  DISPLAY_TREE
+}
