+++ src/main/app/Resources/modules/data/types/choice/components/input.jsx
@@ -11,22 +11,37 @@ import {Radios} from '#/main/app/input/components/radios'
 const ChoiceInput = props => {
   const className = classes('choice-control', props.className, props.size)
 
+  /**
+   * Removes unknown choices to avoid obscure validation error.
+   * (in list configuration, you often get these kind of errors because of deleted / renamed column in the list definition).
+   *
+   * This will also permit to clean the DB over time.
+   */
+  const sanitize = (value) => {
+    let sanitized = value
+    if (props.multiple) {
+      sanitized = (value || []).filter((choice) => !!props.choices[choice])
+    }
+
+    props.onChange(sanitized)
+  }
+
   if (!props.condensed) {
     if (props.multiple) {
       return (
-        <Checkboxes {...props} className={className} />
+        <Checkboxes {...props} className={className} onChange={sanitize} />
       )
     }
 
     if (!props.multiple) {
       return (
-        <Radios {...props} className={className} />
+        <Radios {...props} className={className} onChange={sanitize} />
       )
     }
   }
 
   return (
-    <Select {...props} className={className} />
+    <Select {...props} className={className} onChange={sanitize} />
   )
 }
