+++ src/main/app/Resources/styles/_variables.old.scss
@@ -455,7 +455,7 @@ $data-card-icon-sm-size:      80px;
 $data-card-title-sm-font:     $font-size-base;
 
 $data-card-icon-lg-size:      120px;
-$data-card-title-lg-font:     $font-size-base;
+$data-card-title-lg-font:     $font-size-lg;
 
 // states
 $data-card-color-muted:               var(--#{$prefix}secondary-color) !default;
