+++ src/main/app/Resources/styles/components/_tables.scss
@@ -6,69 +6,14 @@
 
 .table {
     // header
-    > thead > tr > th {
-        position: relative; // for positioning sorting indicator
-        //border-bottom: none !important;
-        font-weight: normal;
-        //padding-top: floor(math.div($table-cell-padding-y, 2));
-        //padding-bottom: floor(math.div($table-cell-padding-y, 2));
+    > thead th {
         font-size: $font-size-sm;
         text-transform: uppercase;
     }
 
     // Fixes vertical align in tables
-    > thead > tr > th,
-    > thead > tr > td,
-    > tbody > tr > th,
-    > tbody > tr > td {
-        vertical-align: middle;
-    }
-
-    > tbody {
-        //@include box-shadow($box-shadow);
-
-        /*th:first-child,
-        td:first-child {
-            border-left: 1px solid $table-border-color;
-        }
-
-        th:last-child,
-        td:last-child {
-            border-right: 1px solid $table-border-color;
-        }
-
-        > tr:last-child th,
-        > tr:last-child td {
-            border-bottom: 1px solid $table-border-color;
-        }*/
-
-        /*> tr {
-            background: $card-bg;
-        }*/
-
-        /*> tr.selected {
-            background-color: $state-info-bg;
-            border-color: #abcfe0; // todo : use a var
-
-            > th,
-            > td {
-                border-color: #abcfe0; // todo : use a var
-            }
-
-            + tr > th,
-            + tr > td {
-                border-top-color: #abcfe0; // todo : use a var
-            }
-        }
-
-        > tr.selected:hover {
-            background-color: #b5d9ea !important; // important is required to override strips color
-        }*/
-    }
-
-    .selected-actions-row > td {
-        padding: 0 !important;
-        border-top: none;
+    > thead {
+        vertical-align: $table-cell-vertical-align; // bootstrap forces it to "bottom"
     }
 
     .list-selected {
@@ -81,28 +26,23 @@
         }
 
         .btn {
-            padding: $table-cell-padding-y - 1 $table-cell-padding-x - 1;  // -1 to absorb btn border
+            padding: $table-cell-padding-y $table-cell-padding-x !important;;
+            border: none !important;;
         }
     }
 }
-
-/*.table-striped {
-    > tbody > tr.selected:nth-of-type(odd) {
-        background-color: #bfe3f4;
-    }
-}*/
-
 
 // -
 // Cells
 // ---
 
 .sorting-cell {
+    position: relative; // for positioning sorting indicator
     padding-right: 24px !important; // arbitrary value to have the sorting indicator centered in the padding. FIXME
     cursor: pointer;
 
     .fa {
-        line-height: 1.42857; // todo find the correct computed line-height
+        line-height: /*1.42857 */$line-height-base; // todo find the correct computed line-height
         position: absolute;
         right: 8px; // same as horizontal padding of the cell
     }
@@ -117,9 +57,11 @@
     width: 50px; // because this cell should only contain a dropdown to gain horizontal space (no buttons bar)
     padding: 0 !important; // move padding on btns
     white-space: nowrap; // avoids buttons to break on multi lines
+    text-transform: none !important;
 
     .btn {
-        padding: $table-cell-padding-y - 1 $table-cell-padding-x - 1;  // -1 to absorb btn border
+        padding: $table-cell-padding-y $table-cell-padding-x !important;
+        border: none !important;;
     }
 
     .dropdown-menu {
@@ -127,8 +69,7 @@
     }
 }
 
-.table-sm/*,
-.table-condensed*/ {
+.table-sm {
     .list-selected {
         .list-selected-label {
             padding: $table-cell-padding-y-sm 0;
@@ -140,6 +81,7 @@
 
     .list-selected .btn,
     .actions-cell .btn {
-        padding: calc($table-cell-padding-y-sm - 1px) calc($table-cell-padding-x-sm - 1px); // -1 to absorb btn border
+        padding: $table-cell-padding-y-sm $table-cell-padding-x-sm  !important;;
+        border: none  !important;;
     }
 }
