+++ src/main/app/Resources/styles/mixins/_lists.scss
@@ -5,13 +5,6 @@
         border-bottom: $list-controls-border-width solid $list-controls-border-color;
     }
 
-    .list-selected {
-        /*border-left: none;
-        border-right: none;*/
-        //border-top: $list-controls-border-width solid $list-controls-border-color; // force border top for table display
-        //border-bottom-color: $list-controls-border-color;
-    }
-
     .list-footer {
         padding: 5px;
         background: $list-controls-bg;
@@ -24,10 +17,6 @@
     }
 
     // customize view modes
-    .data-table {
-        margin-bottom: -($list-controls-border-width) !important; // merge last row and footer borders. important is to override .modal-content > .table
-    }
-
     .data-tree {
         margin-left: 0;
         margin-right: $data-tree-gutter;
@@ -36,29 +25,15 @@
     .data-grid {
         margin-left: $data-grid-gutter;
         margin-right: $data-grid-gutter;
-        /*.data-grid-header,
-        .list-selected-label {
-            margin-left: $data-grid-gutter;
-        }*/
 
         + .list-footer {
             margin-top: $data-grid-gutter;
         }
     }
 
-    /*.data-grid-col {
-        .data-grid-content {
-            margin: (-$data-grid-gutter) 0;
-        }
-    }
+    .data-table {
+        margin-bottom: -($list-controls-border-width) !important; // merge last row and footer borders. important is to override .modal-content > .table
 
-    .data-grid-row {
-        .data-grid-content {
-            margin: (-$data-grid-gutter*.5) 0;
-        }
-    }*/
-
-    .table {
         th:first-child,
         td:first-child {
             border-left: none;
