+++ src/main/core/Resources/translations/platform.en.json
@@ -723,9 +723,9 @@
     "results_per_page": "Results per page",
     "pages": "Pages",
     "current_page": "Page %current% of %pages%",
-    "list_no_results": "No results found.",
+    "list_no_results": "No results",
     "list_search_placeholder": "Search in the list",
-    "list_search_no_results": "No results found. Try to change your search.",
+    "list_search_no_results": "Try to change your search.",
     "list_search_mode": "Search mode",
     "search_full": "Full search",
     "search_unified": "Unified search",
