+++ src/main/core/Resources/translations/platform.fr.json
@@ -743,9 +743,9 @@
     "results_per_page": "Résultats par page",
     "pages": "Pages",
     "current_page": "Page %current% sur %pages%",
-    "list_no_results": "Aucun résultat trouvé.",
+    "list_no_results": "Aucun résultat",
     "list_search_placeholder": "Rechercher dans la liste",
-    "list_search_no_results": "Aucun résultat trouvé. Essayez de changer votre recherche.",
+    "list_search_no_results": "Essayez de changer votre recherche.",
     "list_search_mode": "Mode de recherche",
     "search_full": "Recherche complète",
     "search_unified": "Recherche unifiée",
