+++ src/plugin/forum/Resources/modules/resources/forum/editor/components/editor.jsx
@@ -6,7 +6,7 @@ import {trans} from '#/main/app/intl/translation'
 import {CALLBACK_BUTTON, LINK_BUTTON} from '#/main/app/buttons'
 import {FormData} from '#/main/app/content/form/containers/data'
 import {actions as formActions, selectors as formSelect} from '#/main/app/content/form/store'
-import {constants as listConst} from '#/main/app/content/list/constants'
+import DISPLAY_MODES from '#/main/app/content/list/modes'
 
 import {selectors as resourceSelectors} from '#/main/core/resource/store'
 
@@ -67,8 +67,8 @@ const EditorComponent = (props) =>
             options: {
               noEmpty: true,
               inline: false,
-              choices: Object.keys(listConst.DISPLAY_MODES).reduce((acc, displayMode) => Object.assign(acc, {
-                [displayMode]: listConst.DISPLAY_MODES[displayMode].label
+              choices: Object.keys(DISPLAY_MODES).reduce((acc, displayMode) => Object.assign(acc, {
+                [displayMode]: DISPLAY_MODES[displayMode].label
               }), {})
             }
           }, {
